Fix hundreds of compiler warnings. 2267 remain, using Java 14.
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@20304 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/persistence/BountyDatabase.java
@@ -246,7 +246,7 @@ public class BountyDatabase
 		}
 		
 		String numberString = BountyDatabase.numberByName.get( name );
-		return numberString == null ? 0 : (int) StringUtilities.parseInt( numberString );
+		return numberString == null ? 0 : StringUtilities.parseInt( numberString );
 	}
 
 	public static final String getMonster( String name )
